Create dispatchAction method
In a few different places I needed to get the new state based
on an action, store the action in memory, and store the action
in the DB. I put all of this in a single method called dispatchAction.

Also needed to modify the FE script some for type reasons.